requirements.txt
celery==4.1.0
django==1.11.7
django-extensions==1.9.7
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
psycopg2==2.7.3.2
pydevd==1.1.1
nodeenv==1.2.0
redis==2.10.6
kombu==4.1.0
requests==2.21.0
